# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 供应链金融系统 API文档
  description: 基于FISCO BCOS区块链的供应链金融系统接口文档
  contact:
    name: 开发团队
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Inferred Url
...
paths:
...
  /api/enterprise/register:
    post:
      tags:
      - 企业管理
      summary: 注册企业
      description: 注册新的企业，需要提供企业基本信息
      operationId: registerEnterpriseUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Enterprise'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«Enterprise»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/enterprise/{address}:
    get:
      tags:
      - 企业管理
      summary: 获取企业信息
      description: 根据区块链地址查询企业详细信息
      operationId: getEnterpriseUsingGET
      parameters:
      - name: address
        in: path
        description: 企业区块链地址
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«Enterprise»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/enterprise/{address}/status:
    put:
      tags:
      - 企业管理
      summary: 更新企业状态
      description: 更新企业状态（激活、暂停、拉黑等）
      operationId: updateEnterpriseStatusUsingPUT
      parameters:
      - name: address
        in: path
        description: 企业区块链地址
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
              enum:
              - ACTIVE
              - BLACKLISTED
              - PENDING
              - SUSPENDED
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«Void»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/relations/enterprise/{enterpriseId}:
    get:
      tags:
      - 企业和用户关联查询
      summary: 获取企业及其所有用户
      description: 根据企业ID查询企业详情和所有用户列表
      operationId: getEnterpriseWithUsersUsingGET
      parameters:
      - name: enterpriseId
        in: path
        description: 企业ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«企业用户信息»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/invitation-codes/enterprise/{enterpriseId}:
    get:
      tags:
      - 邀请码管理
      summary: 获取企业邀请码
      description: 查询指定企业的所有邀请码
      operationId: getCodesByEnterpriseUsingGET
      parameters:
      - name: enterpriseId
        in: path
        description: 企业ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«邀请码»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/users/enterprise/{enterpriseId}:
    get:
      tags:
      - 用户管理
      summary: 获取企业用户
      description: 查询指定企业的所有用户
      operationId: getEnterpriseUsersUsingGET
      parameters:
      - name: enterpriseId
        in: path
        description: 企业ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«User»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/users/{userId}/password:
    put:
      tags:
      - 用户管理
      summary: 修改密码
      description: 用户修改自己的密码
      operationId: changePasswordUsingPUT
      parameters:
      - name: userId
        in: path
        description: 用户ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«string»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
components:
  schemas:
...
    ChangePasswordRequest:
      title: ChangePasswordRequest
      required:
      - newPassword
      - oldPassword
      type: object
      properties:
        newPassword:
          type: string
          description: 新密码
        oldPassword:
          type: string
          description: 原密码
...
    Enterprise:
      title: Enterprise
      required:
      - address
      - creditCode
      - name
      - role
      type: object
      properties:
        address:
          type: string
          description: 区块链地址
          example: '0x1234567890abcdef1234567890abcdef12345678'
        creditCode:
          type: string
          description: 统一社会信用代码
          example: 91110000MA001234XY
        creditLimit:
          type: number
          description: 授信额度
          format: bigdecimal
          example: 1000000.0
        creditRating:
          type: integer
          description: 信用评级(0-100)
          format: int32
          example: 75
        enterpriseAddress:
          type: string
          description: 企业地址
          example: 北京市朝阳区
        name:
          type: string
          description: 企业名称
          example: 供应商A
        remarks:
          type: string
          description: 备注信息
          example: 企业审核备注
        role:
          type: string
          description: 企业角色
          example: SUPPLIER
          enum:
          - CORE_ENTERPRISE
          - FINANCIAL_INSTITUTION
          - REGULATOR
          - SUPPLIER
        status:
          type: string
          description: 企业状态
          example: ACTIVE
          enum:
          - ACTIVE
          - BLACKLISTED
          - PENDING
          - SUSPENDED
...
    Result«Enterprise»:
      title: Result«Enterprise»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/Enterprise'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«List«User»»:
      title: Result«List«User»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«List«邀请码»»:
      title: Result«List«邀请码»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/邀请码'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«Void»:
      title: Result«Void»
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«string»:
      title: Result«string»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: string
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«企业用户信息»:
      title: Result«企业用户信息»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/企业用户信息'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    User:
      title: User
      required:
      - status
      - userType
      - username
      type: object
      properties:
        active:
          type: boolean
        avatarUrl:
          type: string
          description: 头像URL
        department:
          type: string
          description: 部门
          example: 财务部
        disabled:
          type: boolean
        email:
          type: string
          description: 电子邮箱
          example: zhangsan@example.com
        enterpriseId:
          type: integer
          description: 所属企业ID
          format: int64
          example: 1
        invitationCode:
          type: string
          description: 注册时使用的邀请码
          example: INV2024011812345678
        locked:
          type: boolean
        phone:
          type: string
          description: 手机号码
          example: '13800138000'
        position:
          type: string
          description: 职位
          example: 财务经理
        realName:
          type: string
          description: 真实姓名
          example: 张三
        registrationRemarks:
          type: string
          description: 注册备注信息
          example: 用户注册申请备注
        status:
          type: string
          description: 用户状态
          enum:
          - ACTIVE
          - DISABLED
          - LOCKED
          - PENDING
        userType:
          type: string
          description: 用户类型
          enum:
          - ADMIN
          - AUDITOR
          - ENTERPRISE_USER
          - OPERATOR
        username:
          type: string
          description: 用户名（登录账号）
          example: zhangsan
    UserInfo:
      title: UserInfo
      type: object
      properties:
        department:
          type: string
          description: 部门
        email:
          type: string
          description: 邮箱
        id:
          type: integer
          description: 用户ID
          format: int64
        loginCount:
          type: integer
          description: 登录次数
          format: int32
        phone:
          type: string
          description: 手机号
        position:
          type: string
          description: 职位
        realName:
          type: string
          description: 真实姓名
        status:
          type: string
          description: 用户状态
        userType:
          type: string
          description: 用户类型
        username:
          type: string
          description: 用户名
...
    企业用户信息:
      title: 企业用户信息
      type: object
      properties:
        address:
          type: string
          description: 企业地址
        creditCode:
          type: string
          description: 统一社会信用代码
        creditLimit:
          type: string
          description: 授信额度
        creditRating:
          type: integer
          description: 信用评级
          format: int32
        enterpriseAddress:
          type: string
          description: 企业地址
        enterpriseId:
          type: integer
          description: 企业ID
          format: int64
        name:
          type: string
          description: 企业名称
        registeredAt:
          type: string
          description: 注册时间
        role:
          type: string
          description: 企业角色
        status:
          type: string
          description: 企业状态
        userCount:
          type: integer
          description: 用户数量
          format: int32
        users:
          type: array
          description: 用户列表
          items:
            $ref: '#/components/schemas/UserInfo'
      description: 企业及其所有用户的详细信息
...
    邀请码:
      title: 邀请码
      required:
      - code
      - createdBy
      - enterpriseId
      - status
      type: object
      properties:
        code:
          type: string
          description: 邀请码（唯一标识）
          example: INV2024011812345678
        createdBy:
          type: string
          description: 创建人
          example: admin
        enterpriseId:
          type: integer
          description: 所属企业ID
          format: int64
          example: 1
        enterpriseName:
          type: string
          description: 企业名称（冗余字段）
          example: 供应商A
        expired:
          type: boolean
        expiresAt:
          type: string
          description: 过期时间
          format: date-time
        maxUses:
          type: integer
          description: 最大使用次数
          format: int32
          example: 100
        maxUsesReached:
          type: boolean
        remarks:
          type: string
          description: 备注信息
          example: 用于招聘财务人员
        status:
          type: string
          description: 邀请码状态
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
        usedCount:
          type: integer
          description: 已使用次数
          format: int32
          example: 5
        valid:
          type: boolean
      description: 企业邀请用户注册的邀请码实体
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header